20240430 commit

+++ lib/providers/plan_shift.dart
@@ -1,6 +1,7 @@
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:miel_work_shift_web/models/organization.dart';
 import 'package:miel_work_shift_web/models/organization_group.dart';
+import 'package:miel_work_shift_web/models/plan_shift.dart';
 import 'package:miel_work_shift_web/services/plan_shift.dart';
 
 class PlanShiftProvider with ChangeNotifier {
@@ -15,7 +16,6 @@ class PlanShiftProvider with ChangeNotifier {
     required bool allDay,
     required bool repeat,
     required String repeatInterval,
-    required int repeatEvery,
     required List<String> repeatWeeks,
     required int alertMinute,
   }) async {
@@ -26,24 +26,27 @@ class PlanShiftProvider with ChangeNotifier {
       return '日時を正しく選択してください';
     }
     try {
-      String id = _planShiftService.id();
-      _planShiftService.create({
-        'id': id,
-        'organizationId': organization.id,
-        'groupId': group?.id ?? '',
-        'userIds': userIds,
-        'startedAt': startedAt,
-        'endedAt': endedAt,
-        'allDay': allDay,
-        'repeat': repeat,
-        'repeatInterval': repeatInterval,
-        'repeatEvery': repeatEvery,
-        'repeatWeeks': repeatWeeks,
-        'alertMinute': alertMinute,
-        'alertedAt': startedAt.subtract(Duration(minutes: alertMinute)),
-        'createdAt': DateTime.now(),
-        'expirationAt': startedAt.add(const Duration(days: 365)),
-      });
+      for (String userId in userIds) {
+        String id = _planShiftService.id();
+        _planShiftService.create({
+          'id': id,
+          'organizationId': organization.id,
+          'groupId': group?.id ?? '',
+          'userId': userId,
+          'startedAt': startedAt,
+          'endedAt': endedAt,
+          'allDay': allDay,
+          'repeat': repeat,
+          'repeatInterval': repeatInterval,
+          'repeatEvery': 1,
+          'repeatWeeks': repeatWeeks,
+          'repeatUntil': null,
+          'alertMinute': alertMinute,
+          'alertedAt': startedAt.subtract(Duration(minutes: alertMinute)),
+          'createdAt': DateTime.now(),
+          'expirationAt': startedAt.add(const Duration(days: 365)),
+        });
+      }
     } catch (e) {
       error = '勤務予定の追加に失敗しました';
     }
@@ -52,37 +55,29 @@ class PlanShiftProvider with ChangeNotifier {
 
   Future<String?> update({
     required String planShiftId,
-    required OrganizationModel? organization,
-    required OrganizationGroupModel? group,
-    required List<String> userIds,
     required DateTime startedAt,
     required DateTime endedAt,
     required bool allDay,
     required bool repeat,
     required String repeatInterval,
-    required int repeatEvery,
     required List<String> repeatWeeks,
     required int alertMinute,
   }) async {
     String? error;
-    if (organization == null) return '勤務予定の編集に失敗しました';
-    if (userIds.isEmpty) return '勤務予定の編集に失敗しました';
     if (startedAt.millisecondsSinceEpoch > endedAt.millisecondsSinceEpoch) {
       return '日時を正しく選択してください';
     }
     try {
       _planShiftService.update({
         'id': planShiftId,
-        'organizationId': organization.id,
-        'groupId': group?.id ?? '',
-        'userIds': userIds,
         'startedAt': startedAt,
         'endedAt': endedAt,
         'allDay': allDay,
         'repeat': repeat,
         'repeatInterval': repeatInterval,
-        'repeatEvery': repeatEvery,
+        'repeatEvery': 1,
         'repeatWeeks': repeatWeeks,
+        'repeatUntil': null,
         'alertMinute': alertMinute,
         'alertedAt': startedAt.subtract(Duration(minutes: alertMinute)),
         'expirationAt': startedAt.add(const Duration(days: 365)),
@@ -94,13 +89,26 @@ class PlanShiftProvider with ChangeNotifier {
   }
 
   Future<String?> delete({
-    required String planShiftId,
+    required PlanShiftModel? planShift,
+    required bool isAllDelete,
   }) async {
     String? error;
+    if (planShift == null) return '勤務予定の削除に失敗しました';
     try {
-      _planShiftService.delete({
-        'id': planShiftId,
-      });
+      if (planShift.repeat) {
+        if (isAllDelete) {
+          _planShiftService.delete({
+            'id': planShift.id,
+          });
+        } else {
+          //更新
+          //登録
+        }
+      } else {
+        _planShiftService.delete({
+          'id': planShift.id,
+        });
+      }
     } catch (e) {
       error = '勤務予定の削除に失敗しました';
     }
